+++ ReverseAndAdd.Test/TestNumberHandler.cs
@@ -10,11 +10,11 @@ namespace ReverseAndAdd.Test
         public void Test_ReverseNumber()
         {
             //Arrange
-            int original = 119;
-            int expected = 911;
+            long original = 119;
+            long expected = 911;
 
             //Act
-            int actual = Logik.NumberHandler.ReverseInput(original);
+            long actual = Logik.NumberHandler.ReverseInput(original);
 
             //Assert
             Assert.AreEqual(expected, actual);
@@ -24,12 +24,12 @@ namespace ReverseAndAdd.Test
         public void Test_AddTwoNumbers()
         {
             //Arrange
-            int firstNumber = 2;
-            int secondNumber = 3;
-            int expected = 5;
+            long firstNumber = 2;
+            long secondNumber = 3;
+            long expected = 5;
 
             //Act
-            int actual = Logik.NumberHandler.AddNumbers(firstNumber, secondNumber);
+            long actual = Logik.NumberHandler.AddNumbers(firstNumber, secondNumber);
 
             //Assert
             Assert.AreEqual(expected, actual);
